+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/huami/zeppos/watches/AmazfitActive3PremiumCoordinator.java
@@ -1,0 +1,94 @@
+/*  Copyright (C) 2026 José Rebelo
+
+    This file is part of Gadgetbridge.
+
+    Gadgetbridge is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as published
+    by the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    Gadgetbridge is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>. */
+package nodomain.freeyourgadget.gadgetbridge.devices.huami.zeppos.watches;
+
+import androidx.annotation.NonNull;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import nodomain.freeyourgadget.gadgetbridge.R;
+import nodomain.freeyourgadget.gadgetbridge.devices.huami.zeppos.ZeppOsCoordinator;
+import nodomain.freeyourgadget.gadgetbridge.impl.GBDevice;
+
+public class AmazfitActive3PremiumCoordinator extends ZeppOsCoordinator {
+    @Override
+    public boolean isExperimental() {
+        return true;
+    }
+
+    @Override
+    public ConnectionType getConnectionType() {
+        return ConnectionType.BOTH;
+    }
+
+    @Override
+    public List<String> getDeviceBluetoothNames() {
+        // never seen, assumptions
+        return Arrays.asList(
+                "Amazfit Active 3 Premium",
+                "Active 3 Premium"
+        );
+    }
+
+    @Override
+    public Set<Integer> getDeviceSources() {
+        return new HashSet<>(Arrays.asList(
+                10944768, // chinese mainland version
+                10944769,
+                10944771,
+                10948867
+        ));
+    }
+
+    @Override
+    public int getDeviceNameResource() {
+        return R.string.devicetype_amazfit_active_3_premium;
+    }
+
+    @Override
+    public boolean mainMenuHasMoreSection() {
+        return false;
+    }
+
+    @Override
+    public boolean supportsControlCenter() {
+        return true;
+    }
+
+    @Override
+    public boolean supportsToDoList() {
+        return true;
+    }
+
+    @Override
+    public boolean supportsWifiHotspot(final GBDevice device) {
+        return true;
+    }
+
+    @Override
+    public boolean supportsBluetoothPhoneCalls(final GBDevice device) {
+        return true;
+    }
+
+    @Override
+    public DeviceKind getDeviceKind(@NonNull GBDevice device) {
+        return DeviceKind.WATCH;
+    }
+}